fix: remove resource store sync to localStorage
Also skip loading ids before loading all info.

These tricks was causing more problems than it was helpful. For instance when switching users or backends.

What we lose now is that the list in the Library is empty for a second
or so.

I think local storage sync was added in the first place back when the backend
relied on Nextcloud for storage and was much slower.

Fixes #55